+++ CHANGELOG.md
@@ -4,3 +4,4 @@
   - Standard repo files
   - Build scripts
     - Initial builds
+  - Argument display and parsing